refactor : Presenter, Contract 패키지 분리
