+++ app/src/main/java/woowacourse/movie/presenter/seat/SeatPresenter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.seat
+package woowacourse.movie.presenter.seat
 
 import woowacourse.movie.domain.model.seat.Seat
 import woowacourse.movie.domain.model.seat.Seats
